pyproject.toml
[project]
name = "urlbot"
version = "0.1.0"
description = "Main script for running the URL Bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "llama-index-core==0.11.8",
  "llama-index-llms-ibm==0.2.0",
  "llama-index-embeddings-ibm==0.2.0",
  "llama-index-readers-web==0.2.2",
  "llama-hub==0.0.79.post1",
  "pydantic==2.9.1",
  "jinja2==3.1.4",
  "fastapi==0.109.2",
  "starlette==0.36.3",
  "huggingface_hub==0.19.4",
  "chromadb==0.4.24",
  "pypdf==4.3.1",
  "trafilatura",
  "python-dotenv",
]

[project.scripts]
urlbot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules"]
include = ["main.py", "config.py"]
